interview-saathi/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
# openai-whisper==20231117
groq==0.9.0
